API/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.4
idna==3.10
psycopg2-binary==2.9.10
requests==2.32.3
setuptools==75.5.0
sqlparse==0.5.1
urllib3>=1.26.5,<2
Pillow
djangorestframework
django-extensions
djangorestframework-simplejwt
channels
django-cors-headers==4.6.0
drf-yasg
msal
django-filter>=22.1
cloudinary==1.32.0
python-decouple
ruff==0.0.287
black==24.3.0
celery
celery[redis]
django-celery-beat
django-celery-results
django-jazzmin
pytest
pytest-django
pytest-cov
drf-yasg
gunicorn
regex